requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
async-generator==1.10
atomicwrites==1.4.1
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
colorama==0.4.6
cryptography==38.0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.4
google-api-core==2.11.0
google-api-python-client==2.68.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
google-oauth==1.0.0
googleapis-common-protos==1.57.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
iniconfig==1.1.1
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
protobuf==4.21.10
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.7.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
toml==0.10.2
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.13
wsproto==1.2.0
xlrd==2.0.1
xmlrunner==1.7.7